+++ how-to/customize-workspace/client/src/framework/shapes/platform-shapes.ts
@@ -1,3 +1,13 @@
+export interface PlatformLocale {
+	/**
+	 * A string with a [BCP 47 language tag](http://tools.ietf.org/html/rfc5646).
+	 *  For the general form and interpretation of the locales argument,
+	 *  see the [`Intl` page](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl#Locale_identification_and_negotiation)
+	 */
+	tag: string;
+	/** An [object](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/Locale/Locale#Parameters) with some or all of options of the locale. */
+	options?: Intl.LocaleOptions;
+}
 export interface PlatformProviderOptions {
 	/** What is the root url of you platform e.g. https://mydomain.com */
 	rootUrl: string;
@@ -13,5 +23,10 @@ export interface PlatformProviderOptions {
 		height?: number;
 		/** the width you wish the window to be */
 		width?: number;
+	};
+	/** internationalization related settings */
+	intl?: {
+		/** If desired specify the preferred locale for this platform */
+		locale?: PlatformLocale;
 	};
 }