+++ how-to/customize-workspace/CHANGELOG.md
@@ -24,6 +24,7 @@
 - Removed `enablePageIntegration` from `homeProvider`, enable/disable the Pages integration instead
 - Added Automation Testing examples
 - Added option of modules receiving a getInteropClient function if they are allowed. Added pattern that a module should listen for the after bootstrap lifecycle event before trying to get an interop client. Added an example to the dev module in modules/composite/developer. There is an analytics implementation that publishes events to an interop/fdc3 channel. This is for dev purposes so you can easily listen to a stream of the events and build a UI (it complements the console analytics module we have). It is only enabled in manifest.fin.json.
+- Added a getLocale option to the module helpers. This returns an object (if specified) that includes the locale tag and optionally the locale options (that can be used to create an Intl.Locale object) specified in the platform provider settings. This provides a way for the platform to pass a preference to dynamically loaded modules (which might be different from the system default language).
 
 ## v9.2